fix : MyPage UI 수정
